extract common function into 'common'
